// src/components/user/types/UserFields.ts
export default interface UserFields {
  firstName: {
    value: string;
    error: string;
  };
  lastName: {
    value: string;
    error: string;
  };
  username: {
    value: string;
    error: string;
  };
  email: {
    value: string;
    error: string;
  };
  password: {
    value: string;
    error: string;
  };
}